+++ crates/plugin_tree_shake/src/module.rs
@@ -65,6 +65,10 @@ pub struct TreeShakeModule {
 
 impl TreeShakeModule {
   pub fn new(module: &Module) -> Self {
+    farmfe_core::farm_profile_function!(format!(
+      "TreeShakeModule::new {:?}",
+      module.id.to_string()
+    ));
     let module_system = module.meta.as_script().module_system.clone();
 
     // 1. generate statement graph
@@ -116,6 +120,10 @@ impl TreeShakeModule {
   }
 
   pub fn used_statements(&self) -> HashMap<StatementId, HashSet<String>> {
+    farmfe_core::farm_profile_function!(format!(
+      "used_statements {:?}",
+      self.module_id.to_string()
+    ));
     // 1. get used exports
     let used_exports_idents = self.used_exports_idents();
     let mut stmt_used_idents_map = HashMap::new();
@@ -127,34 +135,56 @@ impl TreeShakeModule {
       used_idents.insert(used_ident);
     }
 
-    for stmt in self.stmt_graph.stmts() {
-      if stmt.is_self_executed {
-        stmt_used_idents_map
-          .entry(stmt.id)
-          .or_insert(HashSet::new());
+    {
+      farmfe_core::farm_profile_scope!(format!(
+        "analyze self executed stmts {:?}",
+        self.module_id.to_string()
+      ));
+      for stmt in self.stmt_graph.stmts() {
+        if stmt.is_self_executed {
+          stmt_used_idents_map
+            .entry(stmt.id)
+            .or_insert(HashSet::new());
 
-        stmt.used_idents.iter().for_each(|used_ident| {
-          // find the defined ident
-          for stmt_inner in self.stmt_graph.stmts() {
-            if stmt_inner.id == stmt.id {
-              continue;
-            }
+          let dep_stmts = self.stmt_graph.dependencies(&stmt.id);
 
-            if stmt_inner
-              .defined_idents_map
-              .contains_key(&used_ident.to_string())
-              || stmt_inner
-                .defined_idents
-                .iter()
-                .any(|ident| ident.to_string() == used_ident.to_string())
-            {
-              let used_idents = stmt_used_idents_map
-                .entry(stmt_inner.id)
-                .or_insert(HashSet::new());
-              used_idents.insert(UsedIdent::SwcIdent(used_ident.clone()));
-            }
+          for (dep_stmt, referred_idents) in dep_stmts {
+            let used_idents = stmt_used_idents_map
+              .entry(dep_stmt.id)
+              .or_insert(HashSet::new());
+            used_idents.extend(
+              referred_idents
+                .into_iter()
+                .map(|ident| UsedIdent::SwcIdent(ident)),
+            );
           }
-        });
+          // stmt.used_idents.iter().for_each(|used_ident| {
+          //   // find the defined ident
+          //   stmt_used_idents_map
+          //   .entry(stmt.id)
+          //   .or_insert(HashSet::new());
+
+          //   for stmt_inner in self.stmt_graph.stmts() {
+          //     if stmt_inner.id == stmt.id {
+          //       continue;
+          //     }
+
+          //     if stmt_inner
+          //       .defined_idents_map
+          //       .contains_key(&used_ident.to_string())
+          //       || stmt_inner
+          //         .defined_idents
+          //         .iter()
+          //         .any(|ident| ident.to_string() == used_ident.to_string())
+          //     {
+          //       let used_idents = stmt_used_idents_map
+          //         .entry(stmt_inner.id)
+          //         .or_insert(HashSet::new());
+          //       used_idents.insert(UsedIdent::SwcIdent(used_ident.clone()));
+          //     }
+          //   }
+          // });
+        }
       }
     }
 
@@ -167,6 +197,10 @@ impl TreeShakeModule {
   }
 
   pub fn used_exports_idents(&self) -> Vec<(UsedIdent, StatementId)> {
+    farmfe_core::farm_profile_function!(format!(
+      "used_exports_idents {:?}",
+      self.module_id.to_string()
+    ));
     match &self.used_exports {
       UsedExports::All => {
         // all exported identifiers are used