+++ examples/lib-watch/main.ts
@@ -1,0 +1,2 @@
+// Compare this snippet from examples\lib-watch\main.ts:
+export const watch = 'im watch command';