+++ packages/core/src/index.ts
@@ -11,6 +11,7 @@ import compression from 'koa-compress';
 import Koa, { Context } from 'koa';
 import { Compiler } from './compiler/index.js';
 import {
+  normalizePublicDir,
   normalizeUserCompilationConfig,
   resolveUserConfig,
   UserConfig
@@ -20,6 +21,9 @@ import { DevServer } from './server/index.js';
 import { FileWatcher } from './watcher/index.js';
 import type { FarmCLIOptions } from './config/types.js';
 import { Config } from '../binding/index.js';
+import { compilerHandler } from './utils/build.js';
+import { existsSync } from 'node:fs';
+import fse from 'fs-extra';
 
 export async function start(
   options: FarmCLIOptions & UserConfig
@@ -49,8 +53,8 @@ export async function start(
       process.exit(1);
     }
 
-    const fileWatcher = new FileWatcher(userConfig.root, devServer.config.hmr);
-    fileWatcher.watch(devServer, {});
+    const fileWatcher = new FileWatcher(devServer, normalizedConfig);
+    fileWatcher.watch();
   }
 }
 
@@ -58,27 +62,23 @@ export async function build(
   options: FarmCLIOptions & UserConfig
 ): Promise<void> {
   const logger = options.logger ?? new DefaultLogger();
+
   const userConfig: UserConfig = await resolveUserConfig(options, logger);
   const normalizedConfig = await normalizeUserCompilationConfig(
     userConfig,
     'production'
   );
 
-  const start = Date.now();
-  const compiler = new Compiler(normalizedConfig);
-  compiler.removeOutputPathDir();
-  if (userConfig.compilation?.watch) {
-    createFileWatcher(userConfig.root, compiler, normalizedConfig);
-  } else {
-    await compiler.compile();
-    compiler.writeResourcesToDisk();
-    logger.info(
-      `⚡️ Build completed in ${chalk.green(
-        `${Date.now() - start}ms`
-      )}! Resources emitted to ${chalk.green(
-        normalizedConfig.config.output.path
-      )}.`
-    );
+  await createBundleHandler(normalizedConfig);
+
+  // copy resources under publicDir to output.path
+  const absPublicDirPath = normalizePublicDir(
+    normalizedConfig.config.output.path,
+    options.publicDir
+  );
+
+  if (existsSync(absPublicDirPath)) {
+    fse.copySync(absPublicDirPath, normalizedConfig.config.output.path);
   }
 }
 
@@ -143,19 +143,23 @@ export async function watch(
     userConfig,
     'production'
   );
-  const compiler = new Compiler(normalizedConfig);
-  createFileWatcher(userConfig.root, compiler, normalizedConfig);
+
+  createBundleHandler(normalizedConfig, true);
 }
 
-export function createFileWatcher(
-  watcherDirPath: string,
-  compiler: Compiler,
-  normalizedConfig: Config
+export async function createBundleHandler(
+  normalizedConfig: Config,
+  watchMode = false
 ) {
-  const outDir = normalizedConfig.config.output.path;
-  const outDirRegex = new RegExp(`^.*${outDir}.*$`);
-  const fileWatcher = new FileWatcher(watcherDirPath, {
-    ignores: ['**/{.git,node_modules}/**', outDirRegex]
-  });
-  fileWatcher.watch(compiler, normalizedConfig);
+  const compiler = new Compiler(normalizedConfig);
+  await compilerHandler(async () => {
+    compiler.removeOutputPathDir();
+    await compiler.compile();
+    compiler.writeResourcesToDisk();
+  }, normalizedConfig);
+
+  if (normalizedConfig.config?.watch || watchMode) {
+    const watcher = new FileWatcher(compiler, normalizedConfig);
+    watcher.watch();
+  }
 }